+++ src/main/java/madsciencemod/cards/Defend_Bronze.java
@@ -9,6 +9,9 @@ import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 
+import basemod.helpers.BaseModTags;
+import basemod.helpers.CardTags;
+
 public class Defend_Bronze extends AbstractMadScienceCard {
     public static final String ID = "Defend_Bronze";
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
@@ -24,6 +27,7 @@ public class Defend_Bronze extends AbstractMadScienceCard {
     public Defend_Bronze() {
         super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
         this.baseBlock = BLOCK_AMT;
+        CardTags.addTags(this, BaseModTags.BASIC_DEFEND);
     }
 
     @Override