+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.2.2
 * Fixed: Infinite Journal Upgrades Removed Upon Reloading (#12)
+* Fixed: Patch BaseMod upgrade code to check for InfiniteJournal when upgrading cards.
 
 ### v0.2.1
 * Compatibility with ModTheSpire v3.0.0